refactored testing setup
